Export rendertree to achieve go install compatibility
